+++ app/includes/content/data/javascripts.php
@@ -3,6 +3,7 @@
 <script src="js/vendor/jquery-scrollto.js"></script>
 <script src="js/vendor/jquery-history.js"></script>
 <script src="js/vendor/bootstrap.min.js"></script>
+<script src="js/vendor/watch.js"></script>
 <script src="js/plugins.js"></script>
 <script src="js/main.js"></script>
 <script type="text/javascript">