+++ app/includes/layout/progress.php
@@ -1,0 +1,16 @@
+<html>
+	<head>
+		<link rel="stylesheet" href="../../../css/se-dspcentral.css">
+	</head>
+	<body>
+		<div id="progress-display">
+		    <div class="bar">
+		        <span class="bar-fill" id="progressBar"><span class="bar-fill-text" id="pt"></span></span>
+		    </div>
+		    <div id="uploads" class="uploads">
+		        <h3 id="status"></h3>
+		        <p id="loaded_n_total"></p>
+		    </div>
+		</div>
+	</body>
+</html>